Add context.rs back
