+++ pyqir/src/lib.rs
@@ -14,6 +14,7 @@ extern crate llvm_sys_130 as llvm_sys;
 extern crate llvm_sys_140 as llvm_sys;
 
 mod builder;
+mod context;
 mod evaluator;
 mod instructions;
 mod module;
@@ -21,27 +22,3 @@ mod python;
 mod qis;
 mod types;
 mod values;
-
-use pyo3::prelude::*;
-use std::ops::Deref;
-
-/// The context owns global state needed by most LLVM objects.
-#[pyclass]
-#[derive(Eq, PartialEq)]
-pub(crate) struct Context(inkwell::context::Context);
-
-#[pymethods]
-impl Context {
-    #[new]
-    pub(crate) fn new() -> Self {
-        Self(inkwell::context::Context::create())
-    }
-}
-
-impl Deref for Context {
-    type Target = inkwell::context::Context;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
